+++ tools/runners/run-mongo.js
@@ -22,7 +22,7 @@ var runMongoShell = function (url) {
   var auth = mongoUrl.auth && mongoUrl.auth.split(':');
   var ssl = require('querystring').parse(mongoUrl.query).ssl === "true";
 
-  var args = ['--quiet'];
+  var args = [];
   if (ssl) {
     args.push('--ssl');
   }
@@ -61,6 +61,14 @@ function spawnMongod(mongodPath, port, dbPath, replSetName) {
   // Use mmapv1 on 32bit platforms, as our binary doesn't support WT
   if (process.arch === 'ia32') {
     args.push('--storageEngine', 'mmapv1', '--smallfiles');
+  } else if (process.platform !== 'linux') {
+    // MongoDB 4, which we use on 64-bit systems, displays a banner in the
+    // Mongo shell about a free monitoring service, which can be disabled
+    // with this flag. However, the generic Linux build (without SSL; see
+    // MONGO_SSL in scripts/generate-dev-bundle.sh) neither displays the
+    // banner nor supports the flag, so it's safe/important to avoid
+    // passing the flag to mongod on 64-bit linux.
+    args.push('--enableFreeMonitoring', 'off');
   }
 
   return child_process.spawn(mongodPath, args, {