+++ tools/tests/mongo.js
@@ -35,11 +35,20 @@ function testMeteorMongo(appDir) {
   run.waitSecs(15);
 
   var mongoRun = s.run('mongo');
-  run.waitSecs(15);
+  mongoRun.waitSecs(15);
+
+  // Make sure we match the DB version that's printed as part of the
+  // non-quiet shell startup text, so that we don't confuse it with the
+  // output of the db.version() command below.
+  mongoRun.match(/MongoDB server version: 4\.\d+\.\d+/);
+
+  // Make sure the shell does not display the banner about Mongo's free
+  // monitoring service.
+  mongoRun.forbidAll("free cloud-based monitoring service");
+
   // Note: when mongo shell's input is not a tty, there is no prompt.
   mongoRun.write('db.version()\n');
   mongoRun.match(/4\.\d+\.\d+/);
-  run.waitSecs(5);
   mongoRun.stop();
 
   run.stop();